% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{somar}
\alias{somar}
\title{Soma}
\usage{
somar(x, y)
}
\arguments{
\item{x}{\itemize{
\item numero int
}}

\item{y}{\itemize{
\item numero int
}}
}
\description{
Esta função calcula a soma de dois valores
}
